refactor: split memory_store into types/impl/records subfiles
